General: No need to send the current username for several functions. All necessary user information can be obtained via the UserManager.getConnectionBySocket function.
Therefore some changes we needed in some files.

Server/RightManager.js: In loadDefaultRoles, a minor change was needed in the case that an object had no role.

1.rightManager.js and 1.rightmanagerDialog.js: Introduced a readonly variable that disables non manager the possiblity to change something in the right manager.

Client/UserManager.js: Introduced a prepareRoles function that checks whether we a manager or other user is accessing the roles. If the user is no manager, then only those roles will be shown, where the current user is in.